Refactor the compiler plugin directory structure
